backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asyncpg
attrs
certifi
charset-normalizer
click
dataclasses-json
distro
fastapi
filelock
frozenlist
fsspec
groq
h11
hf-xet
httpcore
httpx
httpx-sse
huggingface-hub
idna
Jinja2
joblib
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-postgres
langchain-text-splitters
langsmith
MarkupSafe
marshmallow
mpmath
multidict
mypy_extensions
networkx
numpy
orjson
packaging
pgvector
pillow
propcache
psycopg
psycopg-binary
psycopg-pool
psycopg2-binary
pydantic
pydantic-settings
pydantic_core
python-dotenv
PyYAML
regex
requests
requests-toolbelt
safetensors
scikit-learn
scipy
sentence-transformers
sniffio
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-inspect
typing-inspection
typing_extensions
urllib3
uvicorn
yarl
zstandard
